fix: decimal/float stock arithmetic on order create; add serials, transfers, warranty; shadcn init

+++ backend/app/modules/sales/models/order_item.py
@@ -1,7 +1,7 @@
 import uuid
 from datetime import datetime
 
-from sqlalchemy import DateTime, ForeignKey, Index, Integer, Numeric, String, func
+from sqlalchemy import DateTime, ForeignKey, Index, Numeric, String, func
 from sqlalchemy.dialects.postgresql import JSON, UUID
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
@@ -25,7 +25,7 @@ class OrderItem(UUIDPKMixin, Base):
     sku: Mapped[str | None] = mapped_column(String(100), nullable=True)
     variant_attributes: Mapped[dict | None] = mapped_column(JSON, nullable=True)
     unit_price: Mapped[float] = mapped_column(Numeric(12, 2), nullable=False)
-    quantity: Mapped[int] = mapped_column(Integer, nullable=False, default=1)
+    quantity: Mapped[float] = mapped_column(Numeric(12, 3), nullable=False, default=1)
     discount: Mapped[float] = mapped_column(Numeric(12, 2), default=0)
     line_total: Mapped[float] = mapped_column(Numeric(14, 2), nullable=False)
     created_at: Mapped[datetime] = mapped_column(DateTime(timezone=True), server_default=func.now())
